KVServer, ServerCommand, etc.
Here you can find almost everything for server side. :)
In the KVServer.java there is an implementation for the server. Sever is
a separate program that is running as a thread until you stop it. Each
time when the client wants to establish connection with the server,
server makes separate thread for the client and communicates with him in
that thread.  The implementation of that is in ClientConnection.java, in
run method, server is getting data from client, makes serverCommand for
that data, and then executes command. There are two types of
serverCommnads. First one is GetServerCommand, where server needs to
find and give back data to client. Second one is PutServerCommand where
server needs to put data in server storage.
NOTE: I haven't tested anything yet, so there is a big possibility that
there is some problem in the implementation. But when we merge
everything then we can easily test server. ;)