Implement a very simple test client.
The client uses the KVStore API directly (no CLI or commands)
in order to establish a connection with the KV store server,
and write/read a key/value pair.

This client is here only for the purpose of running the simplest
functionality checks and that is why it's found in a separate
temporary branch.