Fixed put command, added Logger
Here I am again! :)
So, now I have fixed everything about the sending commands. The problem
was that we made the put command for server as well as for the client
just for 2 parameters, and we have forgotten the delete command that is
called also with put command but with one parameter (just key). So now
you can test it and see how it is working.
Also I have added loggers for server and client. Therefore you fill find
out how it is working, it is really simple, just need to get logger ant
write a message. You will see where I have added logs so if I missed
something please add it. :) Also now we have two log files (for user and
for server) divided in separate folders.
NOTE: Logger is printing in log files but also in the command line (that
is just for tasting now, I think that it is enough to print info and
errors in log file), so if you want you can delete
logger.addAppender(consoleAppender); in LogSetup,java, and logger will
not print messages at the command line.