Add a convenience class for serializing MetaData to JSON.
In order to transfer the metadata between different nodes in the
system (client-server, server-ECS), the entire metadata is to be
encoded as a JSON object and sent packaged in a KVMessage's value.

This convenience class provides a mechanism to marshal/unmarshal
the intermediate JSON representation from/to a MetaData object
instance.