Move the communication abstractions to the common package.
This code can be used by the ECS when connecting to particular
servers, so there is no need for duplications.

This refactoring moves the code to the common.communication package
so it is accessible by the ECS too.

It also provides a new/alternative constructor which lets clients
specify the maximum size of the responses which should be handled,
making the code less tightly coupled with the particular use case
of having a KV client communicate with a KV server.