Implement a better MetaData version.
This version of the MetaData only keeps the position of each server
in the ring. It provides a method to find a server which is responsible
for a given key, which is also enough for a server to check if it is
responsible for a given key (if the returned server is the one that
received the request, then it is responsible, otherwise
NOT_RESPONSIBLE).

A method to get the predecessor of a particular hash is also provided
in order to be able to find the key range which needs to be migrated
when a new node should be added to the circle.

A ServerNode class represents a single server and provides methods
which return its name (a concatenation of IP:port, as per the
Milestone 3 text) and its hash (using the md5 hash on the server's
name). The equals and hashCode methods are overridden in order to
be able to compare the ServerNode instances found in the circle for
equality.