Implement the client-side for consistent hashing in the KVStore library.
The library now automatically handles resolving a key to a
responsible server node, retrying requests if issued to a wrong server,
and keeping track of available server nodes.

In order to update the available server nodes, it relies on receiving
the metadata updates from servers upon attempting to contact a server
not responsible for the key.