Add node name as an additional field to ServerNode
Server nodes actually are supposed to be identified by a string
name.

It makes it also easier for server code to identify whether they
are responsible for the received request -- comparing their name
to the name of the node returned by MetaData#getResponsibleServer
is easier (and more reliable) than comparing the IP:port information.