Changed Put and Get commands
Now we have special conditions for put and get commands. First we have
to check the server status. If it is possible to write or read data then
we will see if this server contains that key (key is in the server
range). If the key of the message is not in the range, then we are
sending new meta data to the client. Otherwise we are sending message of
success.