Factor out the interface of the in-memory key/value storage.
This way the code is no longer dependent on having a
ConcurrentHashMap, but only an implementation of the ServerStorage
interface.

The InMemoryServerStorage implementation is backed by a
ConcurrentHashMap.

It would be easy to provide an alternative server implementation now
which uses a different ServerStorage instance which supports
persistence (through dependency injection).