Change the KVStore behavior when the contacted node gives no response.
The KVStore now considers the node dead, instead of continuously
retrying (and thus never succeeding to fullfil the request).

Then, since the node is removed from the ring, the library has a
chance to reach a working/alive node which can at least provide
the current metadata or (at best) reply to the request directly.