Add a "garbage collection" task to the server-nodes.
After a node's metadata is updated, a cleanup thread is spawned
which sweeps through all the data found in the server's storage
and removes anything that the server isn't directly responsible
nor a replica.

This is done in order to be more memory efficient in the node.
Even if the data isn't explicitly removed from the datastore,
the node will exhibit correct behavior when facing the clients
(it will not answer requests for keys that it shouldn't, no matter
that it is technically still found in the datastore and not
explicitly removed yet).